feat: move components and assets into lib folder
