+++ .storybook/preview.ts
@@ -1,4 +1,4 @@
-import '../src/styles/tailwind.css';
+import '../src/lib/styles/tailwind.css';
 
 export const parameters = {
   actions: { argTypesRegex: '^on[A-Z].*' },
